feat: move `no_std` EVM errors to a crate
