+++ Cargo.toml
@@ -1,5 +1,6 @@
 [workspace]
 members = [
+    "crates/evm-errors",
     "crates/interface"
 ]
 resolver = "3"
@@ -15,8 +16,8 @@ keywords = ["blockchain", "zksync"]
 categories = ["cryptography"]
 
 [workspace.dependencies]
-alloy-consensus = { version = "1.0.30", default-features = false }
-alloy-primitives = { version = "1.0.30", default-features = false }
-bytes = { version = "1.10.1", default-features = false }
-portable-atomic = { version = "1.11.1", default-features = false }
+alloy-consensus = { version = "1.0.30" }
+alloy-primitives = { version = "1.0.30" }
 serde = { version = "1.0.219" }
+
+zksync_os_evm_errors = { version = "0.0.5", path = "crates/evm-errors" }
