+++ crates/evm-errors/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "zksync_os_evm_errors"
+description = "EVM errors in ZKsync OS"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+keywords.workspace = true
+categories.workspace = true
+
+[dependencies]
+
+[features]
+default = ["std"]
+std = []
