+++ crates/interface/Cargo.toml
@@ -11,26 +11,8 @@ keywords.workspace = true
 categories.workspace = true
 
 [dependencies]
-alloy-consensus = { workspace = true, default-features = false }
-alloy-primitives = { workspace = true, default-features = false }
-bytes = { workspace = true, default-features = false }
-portable-atomic = { version = "1.11.1", default-features = false, optional = true }
-serde = { workspace = true, features = ["derive"], optional = true }
+zksync_os_evm_errors = { workspace = true }
 
-[features]
-default = ["std", "serde"]
-std = [
-    "alloy-consensus/std",
-    "alloy-primitives/std",
-]
-serde = [
-    "alloy-consensus/serde",
-    "alloy-primitives/serde",
-    "dep:serde",
-]
-# Enables support for single core platforms without atomic CAS.
-unsafe-single-core-cas = [
-    "dep:portable-atomic",
-    "bytes/extra-platforms",
-    "portable-atomic/unsafe-assume-single-core"
-]
+alloy-consensus = { workspace = true, features = ["serde"] }
+alloy-primitives = { workspace = true, features = ["serde"] }
+serde = { workspace = true, features = ["derive"] }
