+++ crates/interface/src/lib.rs
@@ -1,7 +1,3 @@
-#![cfg_attr(not(feature = "std"), no_std)]
-
-extern crate alloc;
-
 pub mod error;
 pub mod tracing;
 pub mod traits;
